+++ clownmdemu.h
@@ -8,7 +8,7 @@
 
 void ClownMDEmu_Init(void *state);
 void ClownMDEmu_Deinit(void *state);
-void ClownMDEmu_Iterate(void *state, void (*video_callback)(size_t scanline, void *pixels, size_t screen_width, size_t screen_height));
+void ClownMDEmu_Iterate(void *state, void (*video_callback)(unsigned short scanline, void *pixels, unsigned short screen_width, unsigned short screen_height));
 void ClownMDEmu_UpdateROM(void *state, const unsigned char *rom_buffer, size_t rom_size);
 void ClownMDEmu_SetROMWriteable(void *state, unsigned char rom_writeable);
 void ClownMDEmu_Reset(void *state_void);
